Update multiple-emails.css
Added "box-sizing: border-box;" CSS so the text input device doesn't cause a strange white space gap in the border of the email container. This is due to the "5px padding" adding extra width to the "100%" width of the textbox. The solution is to change the calculation of the width via box-sizing property.